+++ src/Http/Services/OperatorDateIntervalGetter.php
@@ -2,6 +2,7 @@
 
 namespace ForestAdmin\ForestLaravel\Http\Services;
 
+use Log;
 use Carbon\Carbon;
 
 class OperatorDateIntervalGetter {
@@ -143,9 +144,11 @@ class OperatorDateIntervalGetter {
 
         if (preg_match($this->periodsXHoursBefore, $this->value,
           $this->matches)) {
+            Log::info('This is some useful information.');
+            Log::info(Carbon::now()->subDays($this->matches[1][0]));
             return $this->query->{$this->typeWhere}(function ($query) {
-                $query->where($this->field, '<', Carbon::now()
-                         ->subDays($this->matches[1][0]));
+              $query->where($this->field, '<', Carbon::now()
+                     ->subHours($this->matches[1][0]));
             });
         }