+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Added
-- Filters - Add the before x days operator.
+- Filters - Add the before x hours operator.
 
 ## RELEASE 1.1.4 - 2017-05-30
 ### Added